+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0+4
+
+- Added more methods in ValidatorString.
+- More String Extensions.
+- Added DateTimeExtensions
+- Removed unused code.
+
 ## 0.2.0+3
 
 - Added Validators along with their unit + widget tests in the example folder.